Move migration and consistency check config classes to scenario module
